Add triangulation functionality to mesh module
This commit introduces a new 'triangle' submodule in 'mmcore/topo/mesh', including adding a new Cython implemented 'core.pyx' file. The build and clean scripts have also been updated to allow for these changes. In addition, the commit defines several new macros in 'build.py' for the new triangle library, and makes appropriate modifications in 'clean.py' to maintain the cleaning process.

+++ clean.py
@@ -18,15 +18,19 @@ def delete_files(pattern, path):
 def clean_exts(exts):
     for ext in exts:
         for source in ext.sources:
+
             c_file=source.replace('pyx','c')
-            lib = source.replace('pyx', '*.**.so')
-            try:
-                os.remove(  c_file)
-                print(f"{  c_file} has been deleted.")
-            except OSError as e:
-                print(f"Error: {e.filename} - {e.strerror}.")
-            print("/".join(lib.split('/')[:-1]))
-            delete_files( "*.so","./"+"/".join(lib.split('/')[:-1]))
+            if c_file.endswith('triangle.c'):
+                pass
+            else:
+                lib = source.replace('pyx', '*.**.so')
+                try:
+                    os.remove(  c_file)
+                    print(f"{  c_file} has been deleted.")
+                except OSError as e:
+                    print(f"Error: {e.filename} - {e.strerror}.")
+                print("/".join(lib.split('/')[:-1]))
+                delete_files( "*.so","./"+"/".join(lib.split('/')[:-1]))
 
 if __name__ == '__main__':
     from build import extensions
